refactor(pagination,progressbar,stepper): sépare classe et enregistrement du tag (headless)
