+++ packages/core/src/components/pagination/index.ts
@@ -1,0 +1,11 @@
+import { ArPagination } from './pagination.js';
+
+customElements.define('ar-pagination', ArPagination);
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'ar-pagination': ArPagination;
+    }
+}
+
+export { ArPagination };
