+++ packages/core/src/components/progressbar/index.ts
@@ -1,0 +1,11 @@
+import { ArProgressbar } from './progressbar.js';
+
+customElements.define('ar-progressbar', ArProgressbar);
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'ar-progressbar': ArProgressbar;
+    }
+}
+
+export { ArProgressbar };
