+++ packages/core/src/components/stepper-item/index.ts
@@ -1,0 +1,11 @@
+import { ArStepperItem } from './stepper-item.js';
+
+customElements.define('ar-stepper-item', ArStepperItem);
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'ar-stepper-item': ArStepperItem;
+    }
+}
+
+export { ArStepperItem };
