+++ packages/core/src/components/stepper/index.ts
@@ -1,0 +1,11 @@
+import { ArStepper } from './stepper.js';
+
+customElements.define('ar-stepper', ArStepper);
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'ar-stepper': ArStepper;
+    }
+}
+
+export { ArStepper };
