+++ packages/core/src/index.ts
@@ -14,11 +14,15 @@ import './components/breadcrumb/index.js';
 export { ArBreadcrumb } from './components/breadcrumb/breadcrumb.js';
 import './components/breadcrumb-item/index.js';
 export { ArBreadcrumbItem } from './components/breadcrumb-item/breadcrumb-item.js';
+import './components/progressbar/index.js';
 export { ArProgressbar } from './components/progressbar/progressbar.js';
 import './components/spinner/index.js';
 export { ArSpinner } from './components/spinner/spinner.js';
+import './components/pagination/index.js';
 export { ArPagination } from './components/pagination/pagination.js';
+import './components/stepper/index.js';
 export { ArStepper } from './components/stepper/stepper.js';
+import './components/stepper-item/index.js';
 export { ArStepperItem } from './components/stepper-item/stepper-item.js';
 import './components/dialog/index.js';
 export { ArDialog } from './components/dialog/dialog.js';
